Added ResultsRenderer component with client friendly capabilities. Change ws response from Stats or String to ExperimentResponse. Minor: added totalLifetime to Env, near endTime, typed Long instead of Double. Renamed setup in FitnessCalculator.

+++ algen/elastxy-core/src/main/java/org/elastxy/core/stats/AggregatedStats.java
@@ -41,7 +41,7 @@ public class AggregatedStats {
 	public List<String> bestMatches = new ArrayList<String>();
 
 	
-    public transient List<Double> successExecutionTimes = new ArrayList<Double>();
+    public transient List<Long> successExecutionTimes = new ArrayList<Long>();
     
     public AggregatedStats() {
     }
